chore(app): rename components to match their contexts
