+++ app/javascript/components/profile/recommendations.vue
@@ -1,31 +1,22 @@
 <template>
   <div class="profile-recommendations">
-    <users-rectangle
+    <profile-recommendations-users
       :being-fetched="fetchingRecommendations"
       :recommendations="recommendations"
       type="best"
-      :title="$i18n.t('message.profile.recommendedReviewers')"
-    >
-    </users-rectangle>
-    <users-rectangle
+      :title="$i18n.t('message.profile.recommendedReviewers')" />
+    <profile-recommendations-users
       :being-fetched="fetchingRecommendations"
       :recommendations="recommendations"
       type="worst"
-      :title="$i18n.t('message.profile.notRecommendedReviewers')"
-    >
-    </users-rectangle>
+      :title="$i18n.t('message.profile.notRecommendedReviewers')" />
   </div>
 </template>
 
 <script>
 import { mapState } from 'vuex';
 
-import UsersRectangle from './users-rectangle'
-
 export default {
-  components: {
-    UsersRectangle,
-  },
   computed: mapState({
     recommendations: state => state.profile.recommendations,
     fetchingRecommendations: state => state.profile.fetchingRecommendations,
